+++ CHANGELOG.md
@@ -33,8 +33,8 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 
 ### Versions
 
-- Electron 23.2.1 (was 23.2.0)
-- Chromium 110.0.5481.208 (was 110.0.5481.192)
+- Electron 24.0.0-beta.7 (was 23.2.0)
+- Chromium 112.0.5615.39 (was 110.0.5481.192)
 
 ## [9.7.0](https://github.com/Jelmerro/Vieb/compare/9.6.0...9.7.0) - 2023-04-02